requirements.txt
fastapi
uvicorn
streamlit
requests
sqlalchemy
python-jose
pandas
scikit-learn
matplotlib
textblob
tensorflow
